Make tables threading
This works, but so far it is slower than unthreaded (maybe not
too much of a surprise, there is overhead in making threads,
and also in the extra copying needed for thread memory safety.
Also this means I had to introduce unsafe blocks in the code,
something I am not too happy about. Putting on this branch in
case I want to come back to this work.