Add features to setup
Add the ability to specify a relative table file on the command
line to setup. This is useful for setting up a local version of
a ups product.

Also, add a filter when setting up products, such that if the
product has already been setup previously, its entries are removed
from all the environment variables. This keeps all the variables
from continuously growing when reruning the setup command.