Introduce SETUP env variable
In order to be compatible with standard eups lookups, add an
additional env variable of the form SETUP_<PRODUCT> that contains
various information about what was setup and where it was found.
This code is not the prettiest and may go away in the future as
more features get added to reups, and compatibility with eups
can be loosened.

There were associated changes with this change, such as making
versions returned from the database more sane. i.e. a vector
which may be empty instead of a vector of options.