Move database lookups to function calls
Move looking up the database locations for the system db, and the
user db into their own functions. This will allow the code to be
shared better.

Additionally while refactoring the lookups were made more robust.
Paths are now PathBufs and not strings, and the system db lookup
now handles if there are multiple paths specified in the environment
variable by taking the first path found. Both lookups now verify
that the db path found is actually a directory on disk.