Make exiting with a message into it's own macro
Because exiting reups with an error message is a common task, put
the functionality into a reusable macro.

Because of the way rust macros work the cogs file, with the
macro_use flag, must be placed first in module declarations in the
crate lib file.

This commit also updates all the locations that previously printed
a message and exited.