Create DBFiles in the database in threads
When the db is created, spawn worker threads which will to the
work of creating DBFile objects. In the case where DBFiles are
used in delayed read mode, this does not affect performance much
either for good or bad. When the db is created with the option to
preload DBFiles at creation time this makes a very large difference
in the runtime, as the file system reads, and string allocations
can be done concurrently.