Move prep command and pass through help
Move the prep command to its own module. This separates changes to
the command with changes to the binary. This also puts the text of
the prep command into the library, in case anyone wanted to use it
in their own code.

In the process expand the rsetup command to check for the help
command switch, if it is present then don't try and eval the output
else, run the command and evaluate the output.