Implement Declare subcommand, continue refactoring
Implement the necessary database pieces for declaring new products
to the database. This updated the DBImpl interface, the way the
DB fronted interacts with sources, and changes to the existing
back end. There were some additional changes and fixes to support
the BDImpl refactoring, as the two tasks were interwoven.