Remove template on DBImpl
DBImpl was previously templated to allow for different types of
tables. This proved problematic however, as containers were typed
on dyn object of type DBImpl, which meant there could only be one
template type used anyway if object needed to be kept in the same
container. In the future the table struct will be modified to
support the needs for difference sources of info for a table
rather than having different Table types.